requirements.txt
# Core framework
streamlit>=1.30.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Report generation
reportlab>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# LangChain & Groq
langchain>=0.1.0
langchain-groq>=0.0.1

# File format support
openpyxl>=3.1.0      # Excel (.xlsx)
xlrd>=2.0.1          # Excel (.xls)
pyarrow>=14.0.0      # Parquet files

# Extra utilities
pyyaml>=6.0
pillow>=10.0.0       # Image handling for reports
